+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Updated
 
+- Renamed the `source` property to `reference` in the `Scroll Detector` component.
+- The error message for third-party components has been updated.
+
+### Updated
+
 - Updated the `outside click` dependencies.
 
 ## [0.8.5] - 2023-03-01